refactor(test): moved the extensions tests into src dir
